% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversions.R
\name{allele_complement_vec}
\alias{allele_complement_vec}
\title{Allele complement for strand alignment (SNPs only).}
\usage{
allele_complement_vec(alleles)
}
\description{
Allele complement for strand alignment (SNPs only).
}
\keyword{internal}
