% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{build_bf_matrix}
\alias{build_bf_matrix}
\title{Build a named BF matrix from a finemapped locus data frame}
\usage{
build_bf_matrix(locus_df, lbf_cols)
}
\description{
Build a named BF matrix from a finemapped locus data frame
}
\keyword{internal}
