% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_bf.R
\name{build_lbf_matrix}
\alias{build_lbf_matrix}
\title{Build a named BF matrix from locus data}
\usage{
build_lbf_matrix(locus_df, lbf_cols)
}
\description{
Build a named BF matrix from locus data
}
\keyword{internal}
