% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_zoom.R
\name{build_single_locus_plot}
\alias{build_single_locus_plot}
\title{Build a single ggplot2 locus zoom panel for one trait}
\usage{
build_single_locus_plot(
  gwas,
  chr,
  xrange,
  ens_db,
  ancestry,
  index_snp,
  trait_label
)
}
\description{
Build a single ggplot2 locus zoom panel for one trait
}
\keyword{internal}
