% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{calculate_f_statistic}
\alias{calculate_f_statistic}
\title{calculate_f_statistic: calculates the F-statistic from the P-value}
\usage{
calculate_f_statistic(gwas)
}
\arguments{
\item{gwas}{dataframe with P column}
}
\value{
gwas with F_STAT column
}
\description{
calculate_f_statistic: calculates the F-statistic from the P-value
}
