% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{change_column_names}
\alias{change_column_names}
\title{change_column_names: private function that takes a named list of column names
and changes the supplied data frame's column names accordingly}
\usage{
change_column_names(gwas, columns = list(), remove_extra_columns = F)
}
\arguments{
\item{gwas}{dataframe of gwas to standardise column names}

\item{columns}{named list for}

\item{remove_extra_columns}{logical flag on if we are removing extra columns}
}
\value{
gwas with new column names
}
\description{
change_column_names: private function that takes a named list of column names
and changes the supplied data frame's column names accordingly
}
