% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrpos_to_rsid.R
\name{chrpos_to_rsid}
\alias{chrpos_to_rsid}
\title{Chromosome & position data to variant RSID}
\usage{
chrpos_to_rsid(
  dt,
  chr_col,
  pos_col,
  ea_col = NULL,
  nea_col = NULL,
  flip = "allow",
  alt_rsids = FALSE,
  build = "b37_dbsnp156",
  dbsnp_dir = NA_character_,
  parallel_cores = parallel::detectCores(),
  verbose = TRUE
)
}
\arguments{
\item{dt}{a data.frame like object, or file path, with at least columns (chrom\if{html}{\out{<chr>}}, pos\if{html}{\out{<int>}}, ea\if{html}{\out{<chr>}}, nea\if{html}{\out{<chr>}})}

\item{chr_col}{a string column name; chromosome position}

\item{pos_col}{a string column name; base position}

\item{ea_col}{a string column name; effect allele}

\item{nea_col}{a string column name; non effect allele}

\item{flip}{a string, options: "report", "allow", "no_flip"}

\item{alt_rsids}{a logical, whether to return additional alternate RSIDs}

\item{build}{a string, options: "b37_dbsnp156", "b38_dbsnp156" (corresponds to the appropriate data directory)}

\item{dbsnp_dir}{a string file path to the dbSNP .fst file directory - see setup documentation}

\item{parallel_cores}{an integer, the number of cores/workers to set up the \code{future::multisession} with}

\item{verbose}{a logical, runtime reporting}
}
\value{
a tibble with an RSID column, or (if \code{alt_rsids} is \code{TRUE}) a list with
tibbles: \code{data} and \code{alt_rsids}
}
\description{
Chromosome & position data to variant RSID
}
