% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{coloc_analysis}
\alias{coloc_analysis}
\title{run_coloc_analysis takes two already harmonised gwases, and runs coloc on the results}
\usage{
coloc_analysis(
  first_gwas,
  second_gwas,
  exposure_name,
  chr = NA,
  bp = NA,
  range = NA,
  default_n = NA
)
}
\arguments{
\item{first_gwas}{first gwas to be run through coloc.  This is the gwas that results will be based off}

\item{second_gwas}{second gwas to be run through coloc}
}
\value{
tibble of coloc results (h0 - h4)
}
\description{
run_coloc_analysis takes two already harmonised gwases, and runs coloc on the results
}
