% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_bf.R
\name{coloc_bf_bf_for_locus_pair}
\alias{coloc_bf_bf_for_locus_pair}
\title{Run coloc::coloc.bf_bf for one pair of overlapping loci}
\usage{
coloc_bf_bf_for_locus_pair(l1, l2, trait1_name, trait2_name, p1, p2, p12)
}
\description{
Run coloc::coloc.bf_bf for one pair of overlapping loci
}
\keyword{internal}
