% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_zoom.R
\name{compute_ld_r2_for_locus}
\alias{compute_ld_r2_for_locus}
\title{Compute r2 relative to an index SNP using the local LD reference panel}
\usage{
compute_ld_r2_for_locus(rsids, chr, ancestry, index_snp)
}
\description{
Compute r2 relative to an index SNP using the local LD reference panel
}
\keyword{internal}
