% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{convert_beta_to_or}
\alias{convert_beta_to_or}
\title{convert_beta_to_or: Given a BETA and SE, calculates the OR and lower and upper bounds}
\usage{
convert_beta_to_or(gwas)
}
\arguments{
\item{gwas}{dataframe with the following columns: BETA, SE}
}
\value{
gwas with new columns OR, OR_LB, OR_UB
}
\description{
convert_beta_to_or: Given a BETA and SE, calculates the OR and lower and upper bounds
}
