% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{convert_reference_build_via_liftover}
\alias{convert_reference_build_via_liftover}
\title{convert_reference_build_via_liftover: Change reference build of BP marker from allow list of liftOver conversions}
\usage{
convert_reference_build_via_liftover(
  gwas,
  input_reference_build = reference_builds$GRCh37,
  output_reference_build = reference_builds$GRCh37,
  output_file
)
}
\arguments{
\item{gwas}{GWAS (file or dataframe) of standardised GWAS}

\item{input_reference_build}{string reference build, found in reference_builds list}

\item{output_reference_build}{string reference build that GWAS is to change to, found in reference_builds list}

\item{output_file}{optional output file name to save to}
}
\value{
gwas input is altered and returned
}
\description{
convert_reference_build_via_liftover: Change reference build of BP marker from allow list of liftOver conversions
}
