% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{convert_z_score_to_beta}
\alias{convert_z_score_to_beta}
\title{convert_z_score_to_beta
taken from here: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8432599/ under "Correlation of trans-eQTL effects"}
\usage{
convert_z_score_to_beta(gwas)
}
\description{
convert_z_score_to_beta
taken from here: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8432599/ under "Correlation of trans-eQTL effects"
}
