% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estimate_variance}
\alias{estimate_variance}
\title{Estimate trait variance, internal function}
\usage{
estimate_variance(se, eaf, n)
}
\arguments{
\item{se}{vector of standard errors}

\item{eaf}{vector of MAF (same length as SE)}

\item{n}{sample size}
}
\value{
estimated standard deviation of Y
}
\description{
Estimate trait standard deviation given vectors of variance of coefficients,  MAF and sample size
}
\details{
Estimate is based on var(beta-hat) = var(Y) / (n * var(X))
var(X) = 2\emph{maf}(1-maf)
so we can estimate var(Y) by regressing n*var(X) against 1/var(beta)
}
\keyword{internal}
