% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_comparisons.R
\name{expected_vs_observed_replication}
\alias{expected_vs_observed_replication}
\title{Expected vs observed replication rates}
\usage{
expected_vs_observed_replication(b_disc, se_disc, b_rep, se_rep, alpha = 0.05)
}
\arguments{
\item{b_disc}{vector of clumped incidence hit effects}

\item{se_disc}{the standard errors for incidence effects}

\item{b_rep}{corresponding vector of associations in progression}

\item{se_rep}{standard errors of effects in progression}

\item{alpha}{p-value threshold to check for replication of incidence hits in progression (e.g. try 0.05 or 1e-5)}
}
\description{
For a set of effects that have discovery and replication betas and SEs, this function determines the
extent to which the observed replication rate matches the expected replication rate.
The expected replication rate is based on the assumption that the replication dataset has the same effect sizes
but that the power may be different (e.g. due to allele frequencies or sample sizes) and is reflected in the replication standard errors.
It assesses replication based on concordance of effect direction across discovery and replication, and p-values
surpassing a user-specified p-value threshold.

doi: 10.1038/nature17671
}
