% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{find_finemap_locus_for_region}
\alias{find_finemap_locus_for_region}
\title{Find the finemapped locus file that covers a genomic region}
\usage{
find_finemap_locus_for_region(finemap_dir, chr, bp, range_bp = 1e+06)
}
\description{
Find the finemapped locus file that covers a genomic region
}
\keyword{internal}
