% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_zoom.R
\name{find_index_snp_from_ld_panel}
\alias{find_index_snp_from_ld_panel}
\title{Find the top SNP near a locus center from the LD reference panel}
\usage{
find_index_snp_from_ld_panel(chr, center_bp, window_bp, ancestry)
}
\description{
Find the top SNP near a locus center from the LD reference panel
}
\keyword{internal}
