% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{forest_plot}
\alias{forest_plot}
\title{forest_plot: produce a forest plot from a GWAS file}
\usage{
forest_plot(table, title, output_file, y_column = NA)
}
\arguments{
\item{table}{dataframe with the following columns: BETA, SE}

\item{title}{title of the plot}

\item{output_file}{file to save the plot}

\item{y_column}{column to use for the y-axis}
}
\value{
ggplot2 object
}
\description{
forest_plot: produce a forest plot from a GWAS file
}
