% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{grouped_forest_plot}
\alias{grouped_forest_plot}
\title{grouped_forest_plot: produce a grouped forest plot from a GWAS file}
\usage{
grouped_forest_plot(
  table,
  title,
  group_column,
  output_file,
  p_value_column = NA,
  q_stat_column = NA
)
}
\arguments{
\item{table}{dataframe with the following columns: BETA, SE}

\item{title}{title of the plot}

\item{group_column}{column to use for the group}

\item{output_file}{file to save the plot}

\item{p_value_column}{column to use for the p-value}

\item{q_stat_column}{column to use for the Q-statistic}
}
\value{
ggplot2 object
}
\description{
grouped_forest_plot: produce a grouped forest plot from a GWAS file
}
