% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gwas_region}
\alias{gwas_region}
\title{gwas_region: filter a GWAS file to a region}
\usage{
gwas_region(gwas, chr, bp, range = 5e+05)
}
\arguments{
\item{gwas}{dataframe with the following columns: CHR, BP}

\item{chr}{chromosome}

\item{bp}{base pair position}

\item{range}{range to filter in base pairs}
}
\value{
gwas with filtered rows
}
\description{
gwas_region: filter a GWAS file to a region
}
