% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_bf.R
\name{load_all_finemap_loci}
\alias{load_all_finemap_loci}
\title{Load all finemapped locus files from a set of directories}
\usage{
load_all_finemap_loci(finemap_dirs)
}
\description{
Load all finemapped locus files from a set of directories
}
\keyword{internal}
