% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{load_qtl_gwas_for_mr_result}
\alias{load_qtl_gwas_for_mr_result}
\title{Load the QTL GWAS file for an MR result row}
\usage{
load_qtl_gwas_for_mr_result(mr_result, qtl_dataset)
}
\description{
Load the QTL GWAS file for an MR result row
}
\keyword{internal}
