% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversions.R
\name{map_ref_to_eaf}
\alias{map_ref_to_eaf}
\title{Map effect allele frequency using reference A1 allele frequency (PLINK --freq).}
\usage{
map_ref_to_eaf(ea, oa, ra1, ra2, a1_freq)
}
\description{
Map effect allele frequency using reference A1 allele frequency (PLINK --freq).
}
\keyword{internal}
