% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{miami_plot}
\alias{miami_plot}
\title{miami_plot: produce miami plot of GWAS data from two gwases}
\usage{
miami_plot(
  first_gwas_filename,
  second_gwas_filename,
  miami_plot_file,
  title = "Comparing GWASes",
  chr = NA,
  bp = NA,
  range = NA
)
}
\arguments{
\item{first_gwas_filename}{filename of first GWAS}

\item{second_gwas_filename}{filename of second GWAS}

\item{miami_plot_file}{file to save the miami plot}

\item{title}{title of the plot}

\item{chr}{chromosome to perform miami plot on}

\item{bp}{base pair position to perform miami plot on}

\item{range}{range to filter in base pairs}
}
\description{
miami_plot: produce miami plot of GWAS data from two gwases
}
