% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{parse_bf_bf_result}
\alias{parse_bf_bf_result}
\title{Parse coloc.bf_bf result for QTL analysis}
\usage{
parse_bf_bf_result(result, exposure_name, locus_name, n_snps)
}
\description{
Parse coloc.bf_bf result for QTL analysis
}
\keyword{internal}
