% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_bf.R
\name{parse_pairwise_bf_bf_result}
\alias{parse_pairwise_bf_bf_result}
\title{Parse the result of coloc.bf_bf for a pairwise GWAS comparison}
\usage{
parse_pairwise_bf_bf_result(result, trait1, trait2, locus1, locus2, n_snps)
}
\description{
Parse the result of coloc.bf_bf for a pairwise GWAS comparison
}
\keyword{internal}
