% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversions.R
\name{populate_gene_names}
\alias{populate_gene_names}
\title{populate_gene_names: populate the gene names from the ENSEMBL_ID column}
\usage{
populate_gene_names(gwas)
}
\arguments{
\item{gwas}{dataframe with the following columns: ENSEMBL_ID}
}
\value{
gwas with new column GENE_NAME
}
\description{
populate_gene_names: populate the gene names from the ENSEMBL_ID column
}
