% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversions.R
\name{populate_rsid}
\alias{populate_rsid}
\title{populate_rsid: populate the RSID column from the SNP column}
\usage{
populate_rsid(gwas, option = populate_rsid_options$none)
}
\arguments{
\item{gwas}{dataframe with the following columns: SNP}

\item{option}{option to populate the RSID column}
}
\value{
gwas with new column RSID
}
\description{
populate_rsid: populate the RSID column from the SNP column
}
