% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{run_coloc_on_list_of_datasets}
\alias{run_coloc_on_list_of_datasets}
\title{run_coloc_on_list_of_datasets: run coloc on a list of datasets}
\usage{
run_coloc_on_list_of_datasets(
  first_gwas_list = list(),
  second_gwas_list = list(),
  exposure_name_list = list(),
  chr_list = list(),
  bp_list = list(),
  range = 5e+05,
  default_n = NA,
  output_file
)
}
\arguments{
\item{first_gwas_list}{list of first GWAS files}

\item{second_gwas_list}{list of second GWAS files}

\item{exposure_name_list}{list of exposure names}

\item{chr_list}{list of chromosomes}

\item{bp_list}{list of base pair positions}

\item{range}{range to filter in base pairs}

\item{default_n}{default sample size}

\item{output_file}{file to save the results}
}
\value{
tibble of coloc results
}
\description{
run_coloc_on_list_of_datasets: run coloc on a list of datasets
}
