% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{run_susie_rss_impl}
\alias{run_susie_rss_impl}
\title{Wrapper so tests can mock SuSiE with \code{testthat::local_mocked_bindings()}.}
\usage{
run_susie_rss_impl(z, R, n, L, coverage, min_abs_corr, verbose = FALSE)
}
\description{
Wrapper so tests can mock SuSiE with \code{testthat::local_mocked_bindings()}.
}
\keyword{internal}
