% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{run_system}
\alias{run_system}
\title{Thin wrapper so tests can mock via testthat::local_mocked_bindings().}
\usage{
run_system(
  command,
  wait = TRUE,
  intern = FALSE,
  ignore.stdout = FALSE,
  ignore.stderr = FALSE
)
}
\description{
Thin wrapper so tests can mock via testthat::local_mocked_bindings().
}
\keyword{internal}
