% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot: produce a volcano plot from a GWAS file}
\usage{
volcano_plot(
  results_file,
  title = "Volcano Plot of Results",
  label = "EXPOSURE",
  num_labels = 30,
  output_file,
  p_val = "p.adjusted"
)
}
\arguments{
\item{results_file}{file to save the plot}

\item{title}{title of the plot}

\item{label}{column to use for the label (default "EXPOSURE")}

\item{num_labels}{number of labels to include in the plot (default 30)}

\item{output_file}{file to save the plot}

\item{p_val}{column to use for the p-value (default "p.adjusted")}
}
\value{
ggplot2 object
}
\description{
volcano_plot: produce a volcano plot from a GWAS file
}
