% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_zoom.R
\name{write_completion_file}
\alias{write_completion_file}
\title{Write a completion sentinel file for Snakemake}
\usage{
write_completion_file(completion_file, content = "done")
}
\description{
Write a completion sentinel file for Snakemake
}
\keyword{internal}
