% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{adjust_gwas_data_from_weights_and_save}
\alias{adjust_gwas_data_from_weights_and_save}
\title{adjust_gwas_data_from_weights: Apply a slope correction weight (and SE) to an existing GWAS and save result to disk.
This can be used in conjuction with weights calculated to account for collider bias.
Currently used to work on a GWAS result of Slopehunter.}
\usage{
adjust_gwas_data_from_weights_and_save(
  gwas,
  harmonised_effects,
  collider_bias_type,
  beta,
  se,
  output_file
)
}
\arguments{
\item{gwas:}{gwas that the new collider bias correction will be saved to.  Swaps out BETA, SE, and P}

\item{harmonised_effects:}{a dataframe that includes BETA.incidence, BETA.prognosis, SE.incidence, SE.prognosis}

\item{collider_bias_type:}{string name of adjustment (eg. slopehunter)}

\item{beta:}{number, slope of the correction}

\item{se:}{number, SE of the corrected slope}

\item{output_file:}{name of file that the corrected GWAS will be saved to}
}
\description{
adjust_gwas_data_from_weights: Apply a slope correction weight (and SE) to an existing GWAS and save result to disk.
This can be used in conjuction with weights calculated to account for collider bias.
Currently used to work on a GWAS result of Slopehunter.
}
