% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_comparisons.R
\name{calculate_heterogeneity_scores}
\alias{calculate_heterogeneity_scores}
\title{Test for heterogeneity of effect estimates between GWAS datasets}
\usage{
calculate_heterogeneity_scores(gwases_by_ancestry, heterogeneity_score_file)
}
\arguments{
\item{gwases_by_ancestry}{Named list of data frames, one per GWAS dataset, with at least BETA, SE and SNP columns}

\item{heterogeneity_score_file}{file to save the heterogeneity scores}
}
\value{
List
\itemize{
\item Q = vector of p-values for Cochrane's Q statistic for each SNP
\item Qj = Data frame of per-population outlier q values for each SNP
}
}
\description{
For each SNP this function will provide a Cochran's Q test statistic -
a measure of heterogeneity of effect sizes between datasets. A low p-value means high heterogeneity.
In addition, for every SNP it gives a per-dataset p-value —
interpreted as whether a particular study is an outlier at that SNP.
}
