% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{change_snp_identifiers}
\alias{change_snp_identifiers}
\title{change_snp_identifiers: changes the SNP identifiers of a GWAS to the output reference build}
\usage{
change_snp_identifiers(
  gwas,
  output_file,
  populate_rsid_option = populate_rsid_options$none,
  input_reference_build = reference_builds$GRCh37,
  output_reference_build = reference_builds$GRCh37,
  input_columns = "default"
)
}
\arguments{
\item{gwas:}{filename of gwas to standardise or dataframe of gwas}

\item{output_file:}{file to save standardised gwas}

\item{populate_rsid_option:}{if you want RSID populated or not}

\item{input_reference_build:}{reference build of CHR and BP of data.  Defaults to GRCh37}

\item{output_reference_build:}{reference build of CHR and BP of data.  Defaults to GRCh37}

\item{input_columns:}{column header map used for renaming GWAS:
can be list() of key/value pairs, string of row in predefined_column_maps, or comma separated list of keys=values}

\item{output_columns:}{column header map used for renaming GWAS:
can be list() of key/value pairs, string of row in predefined_column_maps, or comma separated list of keys=values}
}
\value{
modified gwas: saves new gwas in {output_file} if present
}
\description{
change_snp_identifiers: changes the SNP identifiers of a GWAS to the output reference build
}
