% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_comparisons.R
\name{compare_replication_across_all_gwas_permutations}
\alias{compare_replication_across_all_gwas_permutations}
\title{compare_replication_across_all_gwas_permutations}
\usage{
compare_replication_across_all_gwas_permutations(
  gwas_filenames = c(),
  clumped_filenames = c(),
  result_output_file,
  variants_output_file
)
}
\arguments{
\item{gwas_filenames:}{list of filenames of gwases}

\item{clumped_filenames:}{list of clumped list filesnames genetred from gwas, in same order as gwases}

\item{result_output_file:}{data frame of all results of gwas comparisons concatenated}

\item{variants_output_file:}{data frame of every SNP comparison from clumped list concatenated}
}
\description{
takes a vector of gwasse and associated clumps, then runs the expected vs. observed algorithm
}
