% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{conduct_collider_bias_analysis}
\alias{conduct_collider_bias_analysis}
\title{correct_for_collider_bias: do a bunch of collider bias corrections.  Including:
\itemize{
\item SlopeHunter
\item "Corrected Weighted Least Squares" (CWLS, or Dudbridge) Correction
\item MR IVW
}}
\usage{
conduct_collider_bias_analysis(
  incidence_gwas,
  subsequent_gwas,
  clumped_snps_file,
  collider_bias_results_file,
  harmonised_effects_result_file,
  slopehunter_adjusted_file,
  p_value_thresholds = c(0.1, 0.01, 0.001, 1e-05)
)
}
\value{
2 plots: one manhattan plot and one QQ plot (with lambda included)
}
\description{
correct_for_collider_bias: do a bunch of collider bias corrections.  Including:
\itemize{
\item SlopeHunter
\item "Corrected Weighted Least Squares" (CWLS, or Dudbridge) Correction
\item MR IVW
}
}
