% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{convert_or_to_beta}
\alias{convert_or_to_beta}
\title{convert_or_to_beta: Given an OR and lower and upper bounds,
calculates the BETA, and SE.
based on this answer: https://stats.stackexchange.com/a/327684}
\usage{
convert_or_to_beta(gwas)
}
\arguments{
\item{gwas:}{dataframe with the following columns: OR, LB (lower bound), UB (upper bound)}
}
\value{
gwas with new columns BETA and SE
}
\description{
convert_or_to_beta: Given an OR and lower and upper bounds,
calculates the BETA, and SE.
based on this answer: https://stats.stackexchange.com/a/327684
}
