% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_z_to_lbf}
\alias{convert_z_to_lbf}
\title{Generate log Bayes Factor from Z-score}
\usage{
convert_z_to_lbf(
  z,
  se,
  eaf,
  sample_size,
  study_type,
  effect_priors = c(continuous = 0.15, categorical = 0.2)
)
}
\arguments{
\item{z}{Z-score}

\item{se}{Standard error}

\item{eaf}{Allele frequency}

\item{sample_size}{Sample size}

\item{study_type}{Study type}

\item{effect_priors}{Effect priors}
}
\value{
Log Bayes Factor
}
\description{
Generate log Bayes Factor from Z-score
}
