% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{harmonise_gwases}
\alias{harmonise_gwases}
\title{harmonise_gwases: takes a list of gwases, get the SNPs in common
across all datasets arranged to be in the same order}
\usage{
harmonise_gwases(...)
}
\arguments{
\item{:}{elipses of gwases}
}
\value{
: list of harmonised gwases
}
\description{
harmonise_gwases: takes a list of gwases, get the SNPs in common
across all datasets arranged to be in the same order
}
