% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{manhattan_and_qq}
\alias{manhattan_and_qq}
\title{manhattan_and_qq: produce manhattan and qq plot from a GWAS file}
\usage{
manhattan_and_qq(
  gwas_filename,
  manhattan_filename,
  qq_filename,
  include_qq = T
)
}
\arguments{
\item{gwas_filename}{file of a gwas that includes CHR, CP, P, and SNP}

\item{name}{name of plots to be saved (and named as a header in graph)}

\item{save_dir}{defaults to 'scratch/results'}
}
\value{
2 plots: one manhattan plot and one QQ plot (with lambda included)
}
\description{
manhattan_and_qq: produce manhattan and qq plot from a GWAS file
}
