% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{miami_plot}
\alias{miami_plot}
\title{miami_plot: produce miami plot of GWAS data from two gwases}
\usage{
miami_plot(
  first_gwas_filename,
  second_gwas_filename,
  miami_plot_file,
  title = "Comparing GWASes",
  chr = NA,
  bp = NA,
  range = NA
)
}
\arguments{
\item{gwas_dataframe:}{a dataframe that includes CHR, CP, P, and SNP}

\item{name:}{name of plots to be saved (and named as a header in graph)}

\item{save_dir:}{defaults to 'scratch/results'}
}
\description{
miami_plot: produce miami plot of GWAS data from two gwases
}
