% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversions.R
\name{populate_eaf_from_reference_panel}
\alias{populate_eaf_from_reference_panel}
\title{Populate missing EAF from the LD reference panel (1000 Genomes PLINK binary + .frq).}
\usage{
populate_eaf_from_reference_panel(gwas, ancestry, enabled = FALSE)
}
\arguments{
\item{gwas}{Standardised GWAS tibble (CHR, BP, EA, OA; SNP id built after allele ordering).}

\item{ancestry}{One of EUR, EAS, AFR, AMR, SAS matching the reference panel prefix.}

\item{enabled}{If FALSE, returns \code{gwas} unchanged.}
}
\description{
Expects \code{plink --bfile <prefix> --freq --out <prefix>} to exist beside the
\code{.bed/.bim/.fam} files, i.e. \code{<thousand_genomes_dir>/<ANCESTRY>.frq}.
}
