% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{run_susie_for_locus}
\alias{run_susie_for_locus}
\title{Run SuSiE on a single locus}
\usage{
run_susie_for_locus(
  z_scores,
  ld_matrix,
  snp_info,
  n,
  lead_snp,
  max_causal = 10,
  coverage = 0.95,
  min_abs_corr = 0.5
)
}
\arguments{
\item{z_scores}{numeric vector of z-scores}

\item{ld_matrix}{square LD correlation matrix}

\item{snp_info}{data frame with SNP, CHR, BP, RSID columns (same order as z_scores)}

\item{n}{sample size}

\item{lead_snp}{RSID of the lead (clumped) SNP}

\item{max_causal}{max causal signals (L)}

\item{coverage}{credible set coverage}

\item{min_abs_corr}{minimum absolute correlation for CS purity}
}
\value{
tibble with SNP, CHR, BP, RSID, Z, CS, and \code{LBF_1}, \code{LBF_2}, ... per signal
}
\description{
Run SuSiE on a single locus
}
