% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{standardise_gwas}
\alias{standardise_gwas}
\title{standardise_gwas: takes an input gwas, changes headers, standardises allelic input, adds RSID, makes life easier}
\usage{
standardise_gwas(
  gwas,
  output_file,
  N = 0,
  input_format = "default",
  output_format = "default",
  populate_rsid_option = F,
  input_reference_build = reference_builds$GRCh37,
  output_reference_build = reference_builds$GRCh37,
  input_column_map = NULL,
  output_column_map = NULL
)
}
\arguments{
\item{gwas:}{filename of gwas to standardise or dataframe of gwas}

\item{output_file:}{file to save standardised gwas}

\item{input_format:}{type of non-bespoke input format available}

\item{populate_rsid_option:}{if you want RSID populated or not}

\item{input_column_map:}{column header map used for renaming}
}
\value{
modified gwas: saves new gwas in {output_file} if present
}
\description{
standardise_gwas: takes an input gwas, changes headers, standardises allelic input, adds RSID, makes life easier
}
