% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_formatting.R
\name{standardise_gwas}
\alias{standardise_gwas}
\title{standardise_gwas: takes an input gwas, changes headers, standardises allelic input, adds RSID, makes life easier}
\usage{
standardise_gwas(
  gwas,
  output_file,
  N = 0,
  populate_rsid_option = populate_rsid_options$none,
  input_reference_build = reference_builds$GRCh37,
  output_reference_build = reference_builds$GRCh37,
  input_columns = "default",
  output_columns = "default",
  remove_extra_columns = F
)
}
\arguments{
\item{gwas:}{filename of gwas to standardise or dataframe of gwas}

\item{output_file:}{file to save standardised gwas}

\item{N:}{sample size of GWAS (if GWAS has defined N column, defaults to that value)}

\item{populate_rsid_option:}{if you want RSID populated or not}

\item{input_reference_build:}{reference build of CHR and BP of data.  Defaults to GRCh37}

\item{output_reference_build:}{reference build of CHR and BP of data.  Defaults to GRCh37}

\item{input_columns:}{column header map used for renaming GWAS:
can be list() of key/value pairs, string of row in predefined_column_maps, or comma separated list of keys=values}

\item{output_columns:}{column header map used for renaming GWAS:
can be list() of key/value pairs, string of row in predefined_column_maps, or comma separated list of keys=values}

\item{r:}{reference build of CHR and BP of data.  Defaults to GRCh37}
}
\value{
modified gwas: saves new gwas in {output_file} if present
}
\description{
standardise_gwas: takes an input gwas, changes headers, standardises allelic input, adds RSID, makes life easier
}
