% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vroom_snps}
\alias{vroom_snps}
\title{vroom_snps: If you only need to get a handful of SNPs out of a whole GWAS, this saves time and memory
NOTE: only works with data that has been standardised, through \code{standardise_gwas}, or at least a tsv}
\usage{
vroom_snps(gwas_file, snps = c())
}
\description{
vroom_snps: If you only need to get a handful of SNPs out of a whole GWAS, this saves time and memory
NOTE: only works with data that has been standardised, through \code{standardise_gwas}, or at least a tsv
}
